Added a goal for CO2 emissions - still to fix the fact that the model does not reach the goal
This model works with a float in the number of hubs. Which might be a problem (i.e., one can have a non integer number of hubs). The model does not reach the goal: I think the reason is that I am not accounting for the shedded hubs at equilibrium. But how to calculate it?