Started a different model with innovation diffusion
Start of a different model on innovation diffusion. Added a base structure for hubs and funds, and the innovation diffusion is implemented on the company side. Now the two views do not connect at all and the model does not really make sense. Not yet implemented any CO2 emission consideration or similar.