The model works, and there is a connection between all the three views (hubs, demand and company
Based on a goal share of demand that goes through the hub, the system achieves a state where the number of hubs is optimal using the funds based on the gap between the true share of demand and the goal. The innovation diffusion theory is applied to the company side only.