Created excel file with macros for scenarios, found an error in the model
Created excel file with macros: the excel file allows you with one button to select the scenario values that you want for the model.
The error is that the share of demand always goes up and doesn't go down if the share of the companies that use the hub goes down. BAD!!