Fixed "using capacity" rate so that it goes negative when companies are dropping out
Fixed using capacity so now it responds better to the companies behaviour. Before it was wrong because the rate was connected to the goal instead of the gap between the goal and the value. Now it works. Also, fixed efficiency because it was having weird behaviour due to the division in the variable. I have also changed a bit the code in VBA but nothing major (just removed some non necessary lines).