Fixed the company weird behaviour and fixed the excel file
Fixed the company behaviour so that now the decision from the companies side in adding funds is taken depending on how much "monetary value" is left (compared to the maximum they can get.
TO BE DONE! Debug a bit better with synthesim and try different max and minimum values
The excel file now also works.
Need to find better numbers so that the model does not behave so crazy ...